requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7

# ----------------------------------------------------------------------
# مكتبات Kivy و Buildozer لتغليف الأندرويد
# ----------------------------------------------------------------------
# Kivy: إطار عمل الواجهة الرسومية الذي يستخدم الـ Webview
kivy==2.3.0

# plyer: المكتبة التي تسمح بالوصول إلى خدمات الجهاز مثل التخزين
plyer

# python-for-android: أداة البناء التي تتطلبها عملية Buildozer
# يجب وضعها للتأكد من استخدام بيئة البناء الصحيحة
python-for-android